% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterSimmatrix.R
\name{clusterSimmatrix}
\alias{clusterSimmatrix}
\title{Hierarchical Clustering Analysis Function}
\usage{
clusterSimmatrix(simmatrix, dist_mthd = "correlation",
  hclust_mthd = "average", nBoot = 1000, nClusters = FALSE)
}
\arguments{
\item{simmatrix}{symmetric similarity matrix object.}

\item{dist_mthd}{Distance measure to be used. Defaults to "correlation". See help(pvclust).}

\item{hclust_mthd}{Method of clustering. Defaults to "average". See help(pvclust).}

\item{nBoot}{number of bootstrap replications. Defaults to 1000. See help(pvclust).}

\item{nClusters}{number of clusters to run in parallel using 'doParallel'. Defaults to FALSE (non-parallel).}
}
\description{
Function applies hierarchical clustering analysis to similarity matrix, such as one output by 'simmatrixMaker' function. Just a wrapper for pvclust. Output is a pvclust object.
}
