% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CumSumSuite.R
\name{cumsumAtSamplingLocation}
\alias{cumsumAtSamplingLocation}
\title{Cumulative sum at coordinates}
\usage{
cumsumAtSamplingLocation(indivraster, Lat, Lon)
}
\arguments{
\item{indivraster}{RasterLayer representing normalized probability of origin surface}

\item{Lat}{Integer latitude}

\item{Lon}{Integer longitude}
}
\description{
Function estimates cumulative sum of all values in a surface below the value at a specified longitude and latitude.
}
