% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CumSumSuite.R
\name{cumsumbelow}
\alias{cumsumbelow}
\title{Cumulative sum below value}
\usage{
cumsumbelow(vals)
}
\arguments{
\item{vals}{Object of numeric or integer class.}
}
\value{
Returns list of values representing cumulative sum of `val` values less than or equal to the input.
}
\description{
Function that calculates the cumulative sum of values less than or equal to a given value.
}
\examples{
vals <- 1:10
cumsumbelow(vals)

}
