% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isotopeAssignmentModel.R
\name{.assignmentMaker}
\alias{.assignmentMaker}
\title{Make probability-of-origin surfaces}
\usage{
.assignmentMaker(.ID, .isotopeValue, .SD_indv, precip_raster, precip_SD_raster)
}
\arguments{
\item{.ID}{ID value or vector of values (for naming assignment model layers). If missing, will count from 1.}

\item{.isotopeValue}{Isotope precipitation value or vector of values.}

\item{.SD_indv}{error associated with transfer function fit. Value or vector of values. If missing, will assume value of 0.}

\item{precip_raster}{precipitation isoscape raster.}

\item{precip_SD_raster}{precipitation isoscape standard deviation raster.}
}
\description{
Make probability-of-origin surfaces
}
\keyword{internal}
