% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isotopeAssignmentModel.R
\name{.findProductThenNormalize}
\alias{.findProductThenNormalize}
\title{Find Product Then Normalize Function}
\usage{
.findProductThenNormalize(...)
}
\description{
Find Product Then Normalize Function
}
\keyword{internal}
