% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isotopeAssignmentModel.R
\name{.normprodrast}
\alias{.normprodrast}
\title{Find product then normalize function for SpatRasters}
\usage{
.normprodrast(variables)
}
\description{
Find product then normalize function for SpatRasters
}
\keyword{internal}
