% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeOddsSurface.R
\name{makeOddsSurfaces}
\alias{makeOddsSurfaces}
\title{Converts normalized probability surface (e.g. one layer output of isotopeAssignmentModel function) to Odds Ratio surfaces.}
\usage{
makeOddsSurfaces(probabilitySurface, rename = FALSE)
}
\arguments{
\item{probabilitySurface}{Normalized probability surface RasterLayer}

\item{rename}{Character value to append to raster name (e.g. "_odds"). Defaults to FALSE.}
}
\value{
Returns rasterLayer rescaled to Odds Ratio values.
}
\description{
Converts normalized probability surface (e.g. one layer output of isotopeAssignmentModel function) to Odds Ratio surfaces.
}
\examples{
# Generate example probability surfaces.
data(isoscape)
myiso <- rasterFromXYZ(isoscape)
myiso_sd <- rasterFromXYZ(isoscape_sd)
df <- data.frame(ID = c(-100, -80, -50), dD = c(-100, -80, -50), SD_indv = rep(5, 3))
assignmentModels <- isotopeAssignmentModel(ID = df$ID, dD = df$dD, SD_indv = df$SD_indv, precip_raster = myiso, precip_SD_raster = myiso_sd, nClusters = FALSE)

# Convert to odds ratio surfaces.
odds_ratio_surface <- lapply(unstack(assignmentModels), makeOddsSurfaces) \%>\% stack
plot(odds_ratio_surface)

}
