% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeQuantileSurface.R
\name{makeQuantileSurfaces}
\alias{makeQuantileSurfaces}
\title{Converts normalized probability surface (e.g. one layer output of
isotopeAssignmentModel function) to quantile surfaces.}
\usage{
makeQuantileSurfaces(probabilitySurface, rename = FALSE)
}
\arguments{
\item{probabilitySurface}{Normalized probability surface RasterLayer}

\item{rename}{Character value to append to raster name (e.g. "_quantile"). Defaults to FALSE.}
}
\value{
Returns rasterLayer rescaled to quantile values.
}
\description{
Converts normalized probability surface (e.g. one layer output of
isotopeAssignmentModel function) to quantile surfaces.
}
\examples{
# Generate example probability surfaces.
data(isoscape)
myiso <- rasterFromXYZ(isoscape)
myiso_sd <- rasterFromXYZ(isoscape_sd)
df <- data.frame(ID = c(-100, -80, -50), dD = c(-100, -80, -50), SD_indv = rep(5, 3))
assignmentModels <- isotopeAssignmentModel(ID = df$ID, dD = df$dD, SD_indv = df$SD_indv, precip_raster = myiso, precip_SD_raster = myiso_sd, nClusters = FALSE)

# Convert to quantile surfaces.
quantile_surface <-  stack( lapply( unstack(assignmentModels), makeQuantileSurfaces) )
plot(quantile_surface)

}
