% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oddsAtCoordinates.R
\name{oddsAtSamplingLocation}
\alias{oddsAtSamplingLocation}
\title{Odds ratio at coordinates function}
\usage{
oddsAtSamplingLocation(indivraster, Lat, Lon)
}
\arguments{
\item{indivraster}{RasterLayer representing normalized probability of origin surface}

\item{Lat}{Integer latitude}

\item{Lon}{Integer longitude}
}
\description{
Function estimates percentile of each non-NA value within a RasterLayer using the empirical cumulative distribution function, and extracts value at location specified. For more information, see help(ecdf).
}
