% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projectClusters.R
\name{projectClusters}
\alias{projectClusters}
\title{Function to project clusters}
\usage{
projectClusters(assignmentRasters, clustID, simmatrix, namedOutput = "both",
  parallel = FALSE)
}
\arguments{
\item{assignmentRasters}{Rasterstack of isotope assignment models.}

\item{clustID}{vector or integer of cluster assignments for each individual. Must match names and order of 'assignmentRasters'.}

\item{simmatrix}{similarity matrix. Output of 'simmatrixMaker' function. Names must match names of 'assignmentRasters'.}

\item{namedOutput}{Requested output. See 'Details.'}

\item{parallel}{Specify clusters to run in parallel using clusterR package. Defaults to FALSE.}
}
\description{
Evaluate tree and cut. Export cluster identity as a vector (or part of a dataframe).
Project the spatial regions best associated with each cluster.
This function assumes that the levels of 'assignmentRasters' and order of 'clustID' correspond to the same indivID.
Output is raster or rasterstack of cluster spatial projections.
}
\details{
'namedOutput' options:
\itemize{
 \item{"numbered"}{ returns raster of projected clusters numbered by input order.}
 \item{"named"}{ returns raster of projected clusters named by input cluster names.}
 \item{"both}{ returns rasterstack of 'numbered' and 'named' rasters.}
}
}
