% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPrecisionPar.R
\name{getPrecisionPar}
\alias{getPrecisionPar}
\title{Calculate model precision at given threshold values (in parallel).}
\usage{
getPrecisionPar(rasterstack, checkVals, method = FALSE, nCluster = 20)
}
\arguments{
\item{rasterstack}{RasterStack of probability surfaces}

\item{checkVals}{vector of numeric 'threshold' values against which to calculate precision}

\item{method}{is FALSE by default. If character vector, appends a column recording 'method' used.}

\item{nCluster}{is a numeric object specifying how many clusters to form and run in parallel.}
}
\value{
Returns a dataframe of precision values at given threshold.
}
\description{
Function that counts cells (number and proportion) above given values.
}
