% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exampleIsoMonthly.R
\docType{data}
\name{iso_monthly}
\alias{iso_monthly}
\title{Example monthly precipitation isoscape data}
\format{
A data.frame with 43200 rows and 5 columns: \code{x} and \code{y}
  coordinates plus three monthly d2H layers \code{d2h_07}, \code{d2h_08},
  and \code{d2h_09}. Convert to a multi-layer SpatRaster with
  \code{rast(iso_monthly, type = "xyz")}.
}
\usage{
iso_monthly
}
\description{
Three months (July, August, September) of a d2H precipitation isoscape,
provided to demonstrate \code{\link{makeMultiMonthIsoscape}}.
}
\examples{
iso <- rast(iso_monthly, type = "xyz")
}
\references{
Bowen G. J., West J.B., Miller C. C., Zhao L. and Zhang T. (2018) IsoMAP: Isoscapes Modeling, Analysis and Prediction (version 1.0). The IsoMAP Project. http://isomap.org
}
\seealso{
\code{\link{iso_monthly_se}}, \code{\link{makeMultiMonthIsoscape}}
}
\keyword{datasets}
