% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exampleIsoscape.R
\docType{data}
\name{isoscape}
\alias{isoscape}
\title{Example isoscape data}
\format{A dataframe containing 21600 observations, derived from a cropped raster object.}
\source{
\href{https://isomap.rcac.purdue.edu/isomap/fitting/explore?jobseq=66085&caseName=Annual_NoAm_H_1980_2010&userName=caitjcampbell}{IsoMAP Job 66100}
}
\usage{
data(isoscape)
}
\description{
Example isoscape data
}
\references{
Bowen G. J., West J.B., Miller C. C., Zhao L. and Zhang T. (2018) IsoMAP: Isoscapes Modeling, Analysis and Prediction (version 1.0). IsoMAP job 66100, Caitlin J. Campbell. Isoscapes Modeling, Analysis and Prediction (version 1.0). The IsoMAP Project.
(\href{http://isomap.org}{http://isomap.org}).
}
\keyword{datasets}
