% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exampleIsoscape_sd.R
\docType{data}
\name{isoscape_sd}
\alias{isoscape_sd}
\title{Example isoscape standard deviation data}
\format{A dataframe containing 1800 observations, derived from a cropped raster object.}
\source{
\href{https://isomap.rcac.purdue.edu/isomap/DownloadMetadata/download?content=tarball&type=raster&jobseq=66100&username=caitjcampbell&casename=Annual_NoAm_Map_H_1980_2010}{Download IsoMAP Job Map 66100}
}
\usage{
isoscape_sd
}
\description{
Example isoscape standard deviation data
}
\examples{
data(isoscape_sd)
iso_sd <- rasterFromXYZ(isoscape_sd)
}
\references{
Bowen G. J., West J.B., Miller C. C., Zhao L. and Zhang T. (2018) IsoMAP: Isoscapes Modeling, Analysis and Prediction (version 1.0). IsoMAP job 66100, Caitlin J. Campbell. Isoscapes Modeling, Analysis and Prediction (version 1.0). The IsoMAP Project.
(\href{http://isomap.org}{http://isomap.org}).
}
\keyword{datasets}
