% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exampleIsoscape_sd.R
\docType{data}
\name{isoscape_sd}
\alias{isoscape_sd}
\title{Example isoscape standard deviation data}
\format{
A dataframe containing 1800 observations, derived from a cropped raster object.
}
\usage{
isoscape_sd
}
\description{
Example isoscape standard deviation data
}
\examples{
iso_sd <- rasterFromXYZ(isoscape_sd)
}
\references{
Campbell C. J. (2018) IsoMAP job 66100, Isoscapes Modeling, Analysis and Prediction (version 1.0). The IsoMAP Project. http://isomap.org

Bowen G. J., West J.B., Miller C. C., Zhao L. and Zhang T. (2018) IsoMAP: Isoscapes Modeling, Analysis and Prediction (version 1.0). IsoMAP job 66100, Caitlin J. Campbell. Isoscapes Modeling, Analysis and Prediction (version 1.0). The IsoMAP Project. http://isomap.org
}
\keyword{datasets}
