% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simmatrixMaker.R
\name{simmatrixMaker}
\alias{simmatrixMaker}
\title{Generates similarity matrix for rasterStack}
\usage{
simmatrixMaker(assignmentRasters, nClusters = FALSE,
  csvSavePath = FALSE)
}
\arguments{
\item{assignmentRasters}{Input rasterStack}

\item{nClusters}{Clusters to create run in parallel using 'doParallel'. Defaults to FALSE.}

\item{csvSavePath}{Optional savepath to write similarity matrix to csv file. Defaults to FALSE, will not create csv.}
}
\description{
Applies pairwise comparisons of scheoner's D-metric amongst each rasterLayer in a RasterStack to populate a similarity matrix.
}
